apps/authenticity-api/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cytoolz==0.11.2
dnspython==2.2.1
email-validator==1.2.1
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
fastapi==0.77.1
frozenlist==1.3.0
h11==0.13.0
hexbytes==0.2.2
httptools==0.4.0
idna==3.3
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
keras==2.8.0
lru-dict==1.1.7
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.22.3
orjson==3.6.8
parsimonious==0.8.1
Pillow==9.1.0
protobuf==3.20.1
pycodestyle==2.8.0
pycryptodome==3.14.1
pydantic==1.9.0
pyrsistent==0.18.1
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rlp==2.0.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
toolz==0.11.2
typing_extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
varint==1.0.2
watchgod==0.8.2
web3==5.29.0
websockets==9.1
yarl==1.7.2
